+++ src/application/portal/new/toc/toc-content-parser.ts
@@ -1,8 +1,8 @@
 import * as path from "path";
 import * as fs from "fs-extra";
-import { TocGroup, TocCustomPage } from "../types/toc/toc";
+import { TocGroup, TocCustomPage } from "../../../../types/toc/toc";
 
-export class ContentParser {
+export class TocContentParser {
   async parseContentFolder(contentFolderPath: string, workingDirectory: string): Promise<TocGroup[]> {
     if (!await fs.pathExists(contentFolderPath)) {
       return [];
